project_root/requirements.txt
transformers
datasets
pandas
pyyaml
torch
